refactor(runtime): Value splits into the object in hand and the engine that makes one
value.rb now holds only what an instance is — frozen fields read by
name. The class-side factory moves beside it: value/coercion.rb (how a
raw argument becomes a typed value, and the numeric/pattern refusals),
value/admission.rb (closed sets — one_of members and admits:), and
value/invariant_violation.rb. Both modules are extended into Value, so
every caller still says Value.for and Value.build; the three formerly
private_class_method checks stay private through module privacy.
