+++ lib/hecksagain/runtime/value.rb
@@ -1,325 +1,19 @@
 require "json"
 require_relative "../rendering"
+require_relative "value/invariant_violation"
+require_relative "value/coercion"
+require_relative "value/admission"
+
 module Hecksagain
   module Runtime
-    class InvariantViolation < StandardError; end
-
+    # A typed value object in hand: frozen fields, read by name. How one is
+    # MADE — coerced from a raw argument, checked against its declared
+    # numeric types, patterns and closed sets — is the class-side engine in
+    # value/coercion.rb and value/admission.rb, extended here so the door
+    # stays where it always was: `Value.for`, `Value.build`.
     class Value
-      def self.for(aggregate, name, value)
-        attribute = aggregate.attribute(name)
-        return value unless attribute
-
-        for_attribute(aggregate, attribute, value)
-      end
-
-      def self.for_attribute(aggregate, attribute, value)
-        return value if attribute.nil? || value.nil?
-        return hydrate_entity_list(aggregate, attribute, value) if attribute.list?
-        return value unless aggregate.respond_to?(:value_object)
-
-        # THE SET THE ATTRIBUTE NAMES IS CHECKED WHERE THE ATTRIBUTE IS KNOWN.
-        # `build` below sees only the value object, never which attribute asked
-        # for it, so a command argument's `admits:` has to be read here — this
-        # is the door every argument and every head field comes through.
-        #
-        # AFTER coercion, not before: a scalar arrives wrapped in whatever holder
-        # its type names (`{value: "append"}` for an OpName), and checking the
-        # raw payload would be checking the envelope.
-        coerced = aggregate.value_object(attribute.type)
-          .then do |value_object|
-            if value.is_a?(self) && value.type_name == value_object&.hecks_name
-              value
-            elsif value_object
-              build(value_object, fields_for(value_object, attribute.name, value))
-            else
-              value
-            end
-          end
-
-        admit_declared_set(aggregate, attribute, coerced)
-        coerced
-      end
-
-      def self.fields_for(value_object, name, value)
-        return value.transform_keys(&:to_sym) if value.is_a?(Hash)
-        # Mutations may legitimately carry a value object into a differently
-        # named value-object slot with the same declared fields (for example,
-        # PositiveMoney into an Account's Money balance).  Rebuild the target
-        # type from its state; callers at the public boundary still have to
-        # supply an object rather than a scalar.
-        return value.to_h if value.is_a?(self)
-
-        raise TypeMismatch,
-              "#{name} is a #{value_object.hecks_name} — pass its fields as an object, not #{Rendering.describe(value)}"
-      end
-
-      def self.build(value_object, fields)
-        fields = value_object.attributes.each_with_object(fields.transform_keys(&:to_sym)) do |attribute, completed|
-          completed[attribute.name] = attribute.default unless completed.key?(attribute.name) || attribute.default.nil?
-        end
-        admit_member(value_object, fields)
-        check_admitted(value_object, fields)
-        check_numeric_fields(value_object, fields)
-        check_patterns(value_object, fields)
-        value_object.invariants.each do |invariant|
-          next if Bluebook::Expression::Evaluator.call(invariant.canonical, fields)
-
-          raise InvariantViolation,
-                "#{value_object.hecks_name} invariant violated — #{invariant.description} " \
-                "(given #{canonical_fields(fields)})"
-        end
-        new(value_object, fields)
-      end
-
-      # A field declared Integer or Float must ARRIVE as one.
-      #
-      # Without this a String sails into a numeric field and the failure surfaces
-      # later, inside a predicate, as `positive? expects a number, got "three"` —
-      # an EvaluationError, which is NOT a domain refusal. So the runtime broke
-      # where the domain should have said no, and the run contract recorded the
-      # crash beside genuine refusals as though the domain had judged it.
-      #
-      # Checked BEFORE invariants, because an invariant reading a mistyped field
-      # is exactly the thing that used to explode.
-      NUMERIC = { "Integer" => Integer, "Float" => Numeric }.freeze
-      private_class_method def self.check_numeric_fields(value_object, fields)
-        value_object.attributes.each do |attribute|
-          expected = NUMERIC[attribute.type.to_s]
-          next unless expected
-
-          given = fields[attribute.name]
-          next if given.nil? || given.is_a?(expected)
-
-          raise TypeMismatch,
-                "#{value_object.hecks_name}.#{attribute.name} expects #{attribute.type}, got #{Rendering.describe(given)}"
-        end
-      end
-
-      # A field declared with a PATTERN must match it.
-      #
-      # Beside check_numeric_fields and for the same reason : a value that does
-      # not look like what it claims to be is the DOMAIN saying no, and it should
-      # say so here rather than let the wrong shape travel on and surface as a
-      # broken predicate later.
-      #
-      # Which regexes may be written at all is PatternSubset's job, enforced when
-      # the bluebook is declared — so by the time a value arrives here the pattern
-      # is already one both runtimes read the same way, and this is a plain match.
-      # A FIELD OF A VALUE OBJECT MAY NAME A SET TOO.
-      #
-      # Beside check_patterns and for the same reason: `Query::Filter.op` is a
-      # plain String field that admits `Vocabulary::QueryComparator`, and the
-      # value object is what gets built — no attribute passes through
-      # `for_attribute` on the way. Both doors, or the word means one thing on
-      # an argument and nothing on a field.
-      private_class_method def self.check_admitted(value_object, fields)
-        value_object.attributes.each do |attribute|
-          next unless attribute.admits
-
-          admit_declared_set(value_object, attribute, fields[attribute.name])
-        end
-      end
-
-      private_class_method def self.check_patterns(value_object, fields)
-        value_object.attributes.each do |attribute|
-          pattern = attribute.pattern
-          next unless pattern
-
-          given = fields[attribute.name]
-          next if given.nil?
-          next if given.is_a?(String) && Regexp.new(pattern).match?(given)
-
-          raise TypeMismatch,
-                "#{value_object.hecks_name}.#{attribute.name} must match #{pattern}, " \
-                "got #{Rendering.describe(given)}"
-        end
-      end
-
-      def self.hydrate(aggregate, state)
-        state.each_with_object({}) do |(name, value), hydrated|
-          key       = name.to_sym
-          attribute = aggregate.attribute(key)
-          hydrated[key] = attribute ? for_attribute(aggregate, attribute, value) : value
-        end
-      end
-
-      def self.hydrate_entity_list(aggregate, attribute, value)
-        entity = aggregate.entities.find { |candidate| candidate.hecks_name == attribute.type.to_s }
-        return value unless entity
-
-        Array(value).map do |element|
-          next element unless element.is_a?(Hash)
-
-          element.each_with_object({}) do |(name, field_value), hydrated|
-            key = name.to_sym
-            field = entity.attribute(key)
-            hydrated[key] = field ? for_attribute(aggregate, field, field_value) : field_value
-          end
-        end
-      end
-
-      # `Value.identifier` used to live here: hand it a one-field value object
-      # and it opened it, so `identified_by :number` could pass for an identity
-      # and the runtime would guess which field was meant. THAT GUESS IS GONE.
-      # An identity names its field — `identified_by { number.value }` — and the
-      # path is what reaches the scalar. A declaration that names no field is
-      # refused when the bluebook loads, so nothing has to be unwrapped later.
-      #
-      # `scalar` below is a different job and stays: rendering a value object
-      # into a column or a message, where there is no path to consult.
-
-      # `Value.reference_id` lived here, opening a reference to find the id
-      # inside it. A reference IS the id now — refused at the payload gate if it
-      # arrives as anything else — so there is nothing left to open. The comment
-      # it carried said retiring it meant changing how references are STORED ;
-      # that is what happened.
-
-      # A REFERENCE IS AN ID, SO AN OBJECT IS NOT ONE.
-      #
-      # Nothing coerces a reference — `for_attribute` misses on
-      # "Reference<Account>", which is no value object's name, and hands the
-      # argument straight through. That is why the wrapped form went in
-      # unnoticed for as long as it did: there was no place it could be
-      # refused, so whatever the first caller wrote became the shape.
-      #
-      # This is that place. It sits at the payload gate rather than inside
-      # coercion because the sentence names the COMMAND, and `for_attribute`
-      # never learns which command it is serving.
-      #
-      # An Array is deliberately not refused here. A reference is never a list
-      # today, and inventing a rule for a shape the language cannot declare is
-      # how decoration gets written.
-      def self.refuse_object_reference(command, attribute, value)
-        return unless attribute.reference?
-        return unless value.is_a?(Hash) || value.is_a?(self)
-
-        raise TypeMismatch,
-              "#{command.hecks_name} refused — a reference is an id, and " \
-              "#{attribute.name} arrived as an object#{known_by(attribute)}"
-      end
-
-      # "(Account is known by number)" — what to send instead. No article, on
-      # purpose: "an Account" and "a Customer" differ by the target's first
-      # letter, and both runtimes would have to agree on that rule to keep the
-      # refusal byte-identical. Silent when the target is another chapter's,
-      # where this runtime cannot see what it is known by.
-      #
-      # EVERY HEAD, because a caller has to pass every one. This read
-      # `identified_by`, which is the SINGLE head and is nil the moment an
-      # identity has two parts — so a composite target fell through the guard
-      # and the refusal went silent exactly where it had the most to say. A
-      # single-path target reads as it always did.
-      def self.known_by(attribute)
-        heads = Array(attribute.type.resolve&.identity_heads)
-        return "" if heads.empty?
-
-        " (#{attribute.type.target_name} is known by #{heads.join(', ')})"
-      end
-
-      def self.scalar(value)
-        return value unless value.is_a?(self)
-
-        fields = value.to_h
-        return fields.values.first if fields.size == 1
-
-        raise TypeMismatch, "#{value.type_name} has multiple fields and cannot stand in for a scalar"
-      end
-
-      def self.from_identifier(aggregate, attribute, identifier)
-        value_object = aggregate.value_object(attribute.type)
-        return identifier unless value_object
-
-        fields = value_object.attributes
-        return build(value_object, { fields.first.name => identifier }) if fields.size == 1
-
-        raise TypeMismatch, "#{value_object.hecks_name} is a composite identity — an identity must have exactly one field"
-      end
-
-      def self.admit_member(value_object, fields)
-        return if value_object.members.empty?
-
-        discriminant = value_object.attributes.first.name
-        offered      = fields[discriminant]
-        admitted     = value_object.members.map { |member| member[discriminant].to_s }
-        return if admitted.include?(offered.to_s)
-
-        raise InvariantViolation,
-              "#{value_object.hecks_name} admits #{admitted.map(&:inspect).join(', ')} — got #{offered.inspect}"
-      end
-
-      # THE SAME REFUSAL, FOR A SET NAMED SOMEWHERE ELSE.
-      #
-      # `admit_member` above refuses a non-member when the value object BEING
-      # BUILT is itself the closed set — which is the only shape `one_of` can
-      # make, because it SYNTHESISES the set from the values written inline. An
-      # `admits:` attribute is the other direction: the value is an ordinary
-      # String or a plain text holder, and the set it must belong to was
-      # declared once, elsewhere, and is named rather than restated.
-      #
-      # Without this the word was a DECLARATION and nothing more — read by
-      # bin/ir_structs to type a Rust field, read by nobody at the door. A rule
-      # that cannot be the one to refuse is decoration, and this language has
-      # paid for that mistake before.
-      def self.admit_declared_set(owner, attribute, value)
-        return value if attribute.nil? || attribute.admits.nil? || value.nil?
-
-        admitted = admitted_members(owner, attribute)
-        offered  = admitted_scalar(value)
-        return value if admitted.include?(offered.to_s)
-
-        raise InvariantViolation,
-              "#{attribute.name} admits #{attribute.admits} — " \
-              "#{admitted.map(&:inspect).join(', ')} — got #{offered.inspect}"
-      end
-
-      # `Vocabulary::MutationOp` — the aggregate that HOLDS the set, then the
-      # set. Qualified because a closed set is a value object INSIDE an
-      # aggregate, which is exactly why `admits` could not be spelled as a
-      # reference: `reference_to` reaches aggregate heads and nothing below one.
-      #
-      # RESOLVED LATE, like `Reference#resolve` and for the same reason — the
-      # set may be declared further down the file than the attribute that names
-      # it. And REFUSED when it resolves to nothing, also like `Reference`: a
-      # link checked against nothing is worse than no link, because it reads
-      # like a rule.
-      def self.admitted_members(owner, attribute)
-        aggregate_name, set_name = attribute.admits.to_s.split("::", 2)
-        chapter = chapter_of(owner)
-        set     = set_name && chapter&.aggregate(aggregate_name)&.value_object(set_name)
-
-        unless set
-          raise InvariantViolation,
-                "#{attribute.name} admits #{attribute.admits}, which this chapter does not " \
-                "declare — a closed set is named Aggregate::SetName, and it must be one " \
-                "the bluebook actually holds"
-        end
-
-        discriminant = set.attributes.first.name
-        set.members.map { |member| member.to_h[discriminant].to_s }
-      end
-
-      # UP TO THE CHAPTER, from wherever the attribute was declared. A value
-      # object's owner is its aggregate and an aggregate's owner is its chapter,
-      # so the walk stops at the first construct that can answer for an
-      # aggregate BY NAME — which is the chapter, and only the chapter.
-      def self.chapter_of(construct)
-        node = construct
-        node = node.hecks_owner while node && !node.respond_to?(:aggregate) && node.respond_to?(:hecks_owner)
-        node.respond_to?(:aggregate) ? node : nil
-      end
-
-      # An admitted value is a SCALAR however it arrived — as a bare string on a
-      # plain field, or wrapped in the one-field holder its type names.
-      def self.admitted_scalar(value)
-        return value unless value.is_a?(self)
-
-        fields = value.to_h
-        fields.size == 1 ? fields.values.first : value
-      end
-
-      def self.canonical_fields(fields)
-        JSON.generate(fields.sort_by { |name, _| name.to_s }.to_h)
-      end
+      extend Coercion
+      extend Admission
 
       attr_reader :value_object
