+++ lib/hecksagain/runtime/value/invariant_violation.rb
@@ -1,0 +1,5 @@
+module Hecksagain
+  module Runtime
+    class InvariantViolation < StandardError; end
+  end
+end
